Re-arranging files
Starting to separate the ios and android specific files for
cleaner usage.
