+++ src/ffi.rs
@@ -1,0 +1,4 @@
+#[cfg(target_os = "ios")]
+mod ios;
+#[cfg(target_os = "android")]
+mod android;
